+++ src/__init__.py
@@ -1,4 +1,3 @@
 """Nucleo academico PRE-POST."""
 
-__version__ = "0.1.0"
-
+__version__ = "0.2.0"